fix warning C4244 on Windows
'initializing': conversion from 'uint64_t' to 'int',  possible loss of data

+++ Python/optimizer_bytecodes.c
@@ -779,7 +779,7 @@ dummy_func(void) {
             ctx->done = true;
             break;
         }
-        int returning_stacklevel = this_instr->operand1;
+        int returning_stacklevel = (int)this_instr->operand1;
         if (frame_pop(ctx, returning_code, returning_stacklevel)) {
             break;
         }
@@ -797,7 +797,7 @@ dummy_func(void) {
             ctx->done = true;
             break;
         }
-        int returning_stacklevel = this_instr->operand1;
+        int returning_stacklevel = (int)this_instr->operand1;
         if (frame_pop(ctx, returning_code, returning_stacklevel)) {
             break;
         }
@@ -815,7 +815,7 @@ dummy_func(void) {
             ctx->done = true;
             break;
         }
-        int returning_stacklevel = this_instr->operand1;
+        int returning_stacklevel = (int)this_instr->operand1;
         if (frame_pop(ctx, returning_code, returning_stacklevel)) {
             break;
         }
